v1.2 Some visual changes, buttons, panels size and location.
Known bugs fixed, now the Navigation Panel is disabling
when writting a test and enabled when got back to home.

You can't close the app when writting a test except the windows commands.

Background Image changing is now more flexible.

Some possibility of error when creating test has been treated.

The main color of elements are changed to RGB(20,30,70).